RSL-29 fix games statistic and add class for work with storage
