+++ rs-lang/src/components/render/renderLoginPopup.ts
@@ -5,6 +5,7 @@ import { startingPage } from './startingPage';
 import { constants } from '../helpers/constansts';
 import { bookPage } from './bookPage';
 import { state } from '../storage/state';
+import { storageStatistic } from '../games/StorageStatistic';
 class RenderLoginPopup {
   renderLoginForm(container: HTMLElement) {
     const popup = document.createElement('div');
@@ -114,14 +115,7 @@ class RenderLoginPopup {
         storage.tokenExpirationDate = Date.now() + constants.TOKEN_EXPIRE_TIME;
         storage.save();
         state.gamesStatistic = {};
-        const dateGameStatistic = JSON.parse(
-          localStorage.getItem(`${storage.idUser}-gamesStatistic`)!
-        );
-        const dateWordsStatistic = JSON.parse(
-          localStorage.getItem(`${storage.idUser}-wordsStatistic`)!
-        );
-        state.gamesStatistic = { ...dateGameStatistic };
-        state.wordsStatistic = { ...dateWordsStatistic };
+        storageStatistic.getStatisticData();
         startingPage.render();
         await bookPage.getAllDifficult();
         bookPage.getAllLearned();
